[#58] Move: EmailAuth 관련 테스트코드 패키지 이동
