+++ src/main/java/com/clover/youngchat/domain/user/service/UserService.java
@@ -9,6 +9,7 @@ import static com.clover.youngchat.global.exception.ResultCode.NOT_FOUND_USER;
 import static com.clover.youngchat.global.exception.ResultCode.SAME_OLD_PASSWORD;
 import static com.clover.youngchat.global.exception.ResultCode.UNAUTHORIZED_EMAIL;
 
+import com.clover.youngchat.domain.auth.entity.EmailAuth;
 import com.clover.youngchat.domain.user.dto.request.UserEmailAuthCheckReq;
 import com.clover.youngchat.domain.user.dto.request.UserEmailAuthReq;
 import com.clover.youngchat.domain.user.dto.request.UserProfileEditReq;
@@ -20,7 +21,6 @@ import com.clover.youngchat.domain.user.dto.response.UserProfileEditRes;
 import com.clover.youngchat.domain.user.dto.response.UserProfileGetRes;
 import com.clover.youngchat.domain.user.dto.response.UserSignupRes;
 import com.clover.youngchat.domain.user.dto.response.UserUpdatePasswordRes;
-import com.clover.youngchat.domain.user.entity.EmailAuth;
 import com.clover.youngchat.domain.user.entity.User;
 import com.clover.youngchat.domain.user.repository.UserRepository;
 import com.clover.youngchat.global.email.EmailUtil;
